backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
websockets==12.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.9
aiosqlite==0.20.0

# Face recognition
face-recognition==1.3.0
Pillow==10.3.0
numpy==1.26.4

# Cloud storage
boto3==1.34.0
cloudinary==1.40.0

# Notifications
python-telegram-bot==21.3
firebase-admin==6.5.0

# Redis (optional)
redis==5.0.4
